refactor: restructure journal-related test directories
