+++ test/features/journal/ui/widgets/entry_details/delete_icon_widget_test.dart
@@ -11,9 +11,9 @@ import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/services/notification_service.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../../mocks/mocks.dart';
-import '../../../test_data/test_data.dart';
-import '../../../widget_test_utils.dart';
+import '../../../../../mocks/mocks.dart';
+import '../../../../../test_data/test_data.dart';
+import '../../../../../widget_test_utils.dart';
 
 void main() {
   group('DeleteIconWidget', () {
