+++ test/features/journal/ui/widgets/entry_details/entry_datetime_widget_test.dart
@@ -10,9 +10,9 @@ import 'package:lotti/services/editor_state_service.dart';
 import 'package:lotti/services/tags_service.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../../mocks/mocks.dart';
-import '../../../test_data/test_data.dart';
-import '../../../widget_test_utils.dart';
+import '../../../../../mocks/mocks.dart';
+import '../../../../../test_data/test_data.dart';
+import '../../../../../widget_test_utils.dart';
 
 void main() {
   group('EntryDetailFooter', () {
