+++ test/features/journal/ui/widgets/entry_details/entry_detail_footer_test.dart
@@ -15,9 +15,9 @@ import 'package:lotti/services/tags_service.dart';
 import 'package:lotti/services/time_service.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../../mocks/mocks.dart';
-import '../../../test_data/test_data.dart';
-import '../../../widget_test_utils.dart';
+import '../../../../../mocks/mocks.dart';
+import '../../../../../test_data/test_data.dart';
+import '../../../../../widget_test_utils.dart';
 
 void main() {
   group('EntryDetailFooter', () {
